packages/backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# AI / ML
google-genai>=1.0.0

# Image Processing
scikit-image==0.25.0
Pillow==11.1.0
numpy==2.2.0
opencv-python-headless==4.11.0.86

# Utilities
python-dotenv==1.0.1
httpx==0.28.0
websockets==14.1

# Testing
pytest==8.3.0
pytest-asyncio==0.25.0
